+++ R/irks-package.R
@@ -1,0 +1,21 @@
+#' Functions To Develop Credit Risk Models
+#' @name irks
+#' @docType package
+#' @author Joshua Kunst (@@jbkunst)
+#' @importFrom stats as.formula
+NULL
+
+#' irks exported operators and S3 methods
+#'
+#' The following functions are imported and then re-exported
+#' from the irks package to avoid listing the magrittr
+#' as Depends of irks.
+#'
+#' @name irks-exports
+NULL
+
+#' @importFrom magrittr %>%
+#' @name %>%
+#' @export
+#' @rdname irks-exports
+NULL