+++ NEWS.md
@@ -1,3 +1,5 @@
 # irks 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.
+* Added `var_imporance` and `ct_rules` from the `jbkmisc` package
+